// app/model/EventRule.ts
/**
 * AMT - gamification
 * The API specification for the AMT - gamification project. All endpoints are implicitely prefixed by __/api__ for the sake of readability.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * A rule applied on event creation.
 */
export interface EventRule {
    /**
     * The name of the rule.
     */
    name?: string;

    /**
     * The event on which the rule is applied.
     */
    event?: string;

    /**
     * The point scale on which the rule is applied.
     */
    pointScale?: string;

    /**
     * How many points are given on the point scale.
     */
    pointsGiven?: number;

}