// app/model/TriggerRule.ts
/**
 * AMT - gamification
 * The API specification for the AMT - gamification project. All endpoints are implicitely prefixed by __/api__ for the sake of readability.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * A rule applied on point scale change.
 */
export interface TriggerRule {
    /**
     * The name of the rule.
     */
    name?: string;

    /**
     * The badge awarded when rule condition matched.
     */
    badgeAwarded?: string;

    /**
     * The point scale on which the rule is applied.
     */
    pointScale?: string;

    /**
     * The point limit to award the badge.
     */
    limit?: number;

    /**
     * Whether the badge is given above or below point limit.
     */
    aboveLimit?: boolean;

}